+++ src/Misc/Hooks.Ares.cpp
@@ -1,7 +1,5 @@
 #include <BuildingClass.h>
 #include <FootClass.h>
-#include <UnitClass.h>
-#include <CellClass.h>
 
 #include <Utilities/Macro.h>
 
@@ -21,36 +19,4 @@ DEFINE_HOOK(0x44E9FA, BuildingClass_Detach_RestoreAnims, 0x6)
 		return R->Origin() == 0x44E997 ? SkipAnimOne : SkipAnimTwo;
 
 	return 0;
-}
-
-// Ares make chrono miner cannot teleport again after repaired from repair bay for unknown reason
-// This hook trying to fix it
-DEFINE_HOOK(0x741B30, UnitClass_SetDestination_HarvesterFix, 0x7)
-{
-	GET(UnitClass*, pThis, EBP);
-
-	if (pThis->GetCurrentMission() != Mission::Sleep)
-		return 0x741BAA;
-
-	if (const auto pLink = pThis->GetNthLink())
-	{
-		const auto pBuilding = abstract_cast<BuildingClass*>(pLink);
-
-		if (pBuilding && pBuilding->Type->UnitRepair && pThis->Locomotor.GetInterfacePtr())
-		{
-			pThis->Locomotor->Power_On();
-			pThis->QueueMission(Mission::Harvest, false);
-		}
-
-		return 0x741B4A;
-	}
-	else if (const auto pCell = pThis->GetCell())
-	{
-		const auto pBuilding = pCell->GetBuilding();
-
-		if (pBuilding && pBuilding->Type->UnitRepair && pThis->Locomotor.GetInterfacePtr())
-			pThis->Locomotor->Power_On();
-	}
-
-	return 0x741BAA;
 }